Reorganize sync deriver logic
