+++ op-node/rollup/driver/constants.go
@@ -1,0 +1,9 @@
+package driver
+
+import "github.com/ethereum-optimism/optimism/op-node/rollup/sequencing"
+
+// aliases to not disrupt op-conductor code
+var (
+	ErrSequencerAlreadyStarted = sequencing.ErrSequencerAlreadyStarted
+	ErrSequencerAlreadyStopped = sequencing.ErrSequencerAlreadyStopped
+)
