+++ op-node/rollup/driver/sync_deriver.go
@@ -4,199 +4,17 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	gosync "sync"
-	"time"
-
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/log"
 
 	"github.com/ethereum-optimism/optimism/op-node/rollup"
 	"github.com/ethereum-optimism/optimism/op-node/rollup/clsync"
 	"github.com/ethereum-optimism/optimism/op-node/rollup/derive"
 	"github.com/ethereum-optimism/optimism/op-node/rollup/engine"
-	"github.com/ethereum-optimism/optimism/op-node/rollup/sequencing"
 	"github.com/ethereum-optimism/optimism/op-node/rollup/status"
 	"github.com/ethereum-optimism/optimism/op-node/rollup/sync"
 	"github.com/ethereum-optimism/optimism/op-service/eth"
 	"github.com/ethereum-optimism/optimism/op-service/event"
+	"github.com/ethereum/go-ethereum/log"
 )
-
-// Deprecated: use eth.SyncStatus instead.
-type SyncStatus = eth.SyncStatus
-
-type Driver struct {
-	StatusTracker SyncStatusTracker
-	Finalizer     Finalizer
-
-	*SyncDeriver
-
-	sched *StepSchedulingDeriver
-
-	emitter event.Emitter
-	drain   Drain
-
-	// Requests to block the event loop for synchronous execution to avoid reading an inconsistent state
-	stateReq chan chan struct{}
-
-	// Upon receiving a channel in this channel, the derivation pipeline is forced to be reset.
-	// It tells the caller that the reset occurred by closing the passed in channel.
-	forceReset chan chan struct{}
-
-	// Driver config: verifier and sequencer settings.
-	// May not be modified after starting the Driver.
-	driverConfig *Config
-
-	// Interface to signal the L2 block range to sync.
-	altSync AltSync
-
-	sequencer sequencing.SequencerIface
-
-	metrics Metrics
-	log     log.Logger
-
-	wg gosync.WaitGroup
-
-	driverCtx    context.Context
-	driverCancel context.CancelFunc
-}
-
-// Start starts up the state loop.
-// The loop will have been started iff err is not nil.
-func (s *Driver) Start() error {
-	log.Info("Starting driver", "sequencerEnabled", s.driverConfig.SequencerEnabled,
-		"sequencerStopped", s.driverConfig.SequencerStopped, "recoverMode", s.driverConfig.RecoverMode)
-	if s.driverConfig.SequencerEnabled {
-		if s.driverConfig.RecoverMode {
-			log.Warn("sequencer is in recover mode")
-			s.sequencer.SetRecoverMode(true)
-		}
-		if err := s.sequencer.SetMaxSafeLag(s.driverCtx, s.driverConfig.SequencerMaxSafeLag); err != nil {
-			return fmt.Errorf("failed to set sequencer max safe lag: %w", err)
-		}
-		if err := s.sequencer.Init(s.driverCtx, !s.driverConfig.SequencerStopped); err != nil {
-			return fmt.Errorf("persist initial sequencer state: %w", err)
-		}
-	}
-
-	s.wg.Add(1)
-	go s.eventLoop()
-
-	return nil
-}
-
-func (s *Driver) Close() error {
-	s.driverCancel()
-	s.wg.Wait()
-	s.sequencer.Close()
-	return nil
-}
-
-// the eventLoop responds to L1 changes and internal timers to produce L2 blocks.
-func (s *Driver) eventLoop() {
-	defer s.wg.Done()
-	s.log.Info("State loop started")
-	defer s.log.Info("State loop returned")
-
-	defer s.driverCancel()
-
-	// reqStep requests a derivation step nicely, with a delay if this is a reattempt, or not at all if we already scheduled a reattempt.
-	reqStep := func() {
-		s.sched.RequestStep(s.driverCtx, false)
-	}
-
-	// We call reqStep right away to finish syncing to the tip of the chain if we're behind.
-	// reqStep will also be triggered when the L1 head moves forward or if there was a reorg on the
-	// L1 chain that we need to handle.
-	reqStep()
-
-	sequencerTimer := time.NewTimer(0)
-	var sequencerCh <-chan time.Time
-	var prevTime time.Time
-	// planSequencerAction updates the sequencerTimer with the next action, if any.
-	// The sequencerCh is nil (indefinitely blocks on read) if no action needs to be performed,
-	// or set to the timer channel if there is an action scheduled.
-	planSequencerAction := func() {
-		nextAction, ok := s.sequencer.NextAction()
-		if !ok {
-			if sequencerCh != nil {
-				s.log.Info("Sequencer paused until new events")
-			}
-			sequencerCh = nil
-			return
-		}
-		// avoid unnecessary timer resets
-		if nextAction == prevTime {
-			return
-		}
-		prevTime = nextAction
-		sequencerCh = sequencerTimer.C
-		if len(sequencerCh) > 0 { // empty if not already drained before resetting
-			<-sequencerCh
-		}
-		delta := time.Until(nextAction)
-		s.log.Info("Scheduled sequencer action", "delta", delta)
-		sequencerTimer.Reset(delta)
-	}
-
-	// Create a ticker to check if there is a gap in the engine queue. Whenever
-	// there is, we send requests to sync source to retrieve the missing payloads.
-	syncCheckInterval := time.Duration(s.Config.BlockTime) * time.Second * 2
-	altSyncTicker := time.NewTicker(syncCheckInterval)
-	defer altSyncTicker.Stop()
-	lastUnsafeL2 := s.Engine.UnsafeL2Head()
-
-	for {
-		if s.driverCtx.Err() != nil { // don't try to schedule/handle more work when we are closing.
-			return
-		}
-
-		planSequencerAction()
-
-		// If the engine is not ready, or if the L2 head is actively changing, then reset the alt-sync:
-		// there is no need to request L2 blocks when we are syncing already.
-		if head := s.Engine.UnsafeL2Head(); head != lastUnsafeL2 || !s.Derivation.DerivationReady() {
-			lastUnsafeL2 = head
-			altSyncTicker.Reset(syncCheckInterval)
-		}
-
-		select {
-		case <-sequencerCh:
-			s.Emitter.Emit(s.driverCtx, sequencing.SequencerActionEvent{})
-		case <-altSyncTicker.C:
-			// Check if there is a gap in the current unsafe payload queue.
-			ctx, cancel := context.WithTimeout(s.driverCtx, time.Second*2)
-			err := s.checkForGapInUnsafeQueue(ctx)
-			cancel()
-			if err != nil {
-				s.log.Warn("failed to check for unsafe L2 blocks to sync", "err", err)
-			}
-		case <-s.sched.NextDelayedStep():
-			s.sched.AttemptStep(s.driverCtx)
-		case <-s.sched.NextStep():
-			s.sched.AttemptStep(s.driverCtx)
-		case respCh := <-s.stateReq:
-			respCh <- struct{}{}
-		case respCh := <-s.forceReset:
-			s.log.Warn("Derivation pipeline is manually reset")
-			s.Derivation.Reset()
-			s.metrics.RecordPipelineReset()
-			close(respCh)
-		case <-s.drain.Await():
-			if err := s.drain.Drain(); err != nil {
-				if s.driverCtx.Err() != nil {
-					return
-				} else {
-					s.log.Error("unexpected error from event-draining", "err", err)
-					s.Emitter.Emit(s.driverCtx, rollup.CriticalErrorEvent{
-						Err: fmt.Errorf("unexpected error: %w", err),
-					})
-				}
-			}
-		case <-s.driverCtx.Done():
-			return
-		}
-	}
-}
 
 type SyncDeriver struct {
 	// The derivation pipeline is reset whenever we reorg.
@@ -427,90 +245,4 @@ func (s *SyncDeriver) SyncStep() {
 	// to generate new attributes, if no attributes are known already.
 	s.Engine.RequestPendingSafeUpdate(s.Ctx)
 
-}
-
-// ResetDerivationPipeline forces a reset of the derivation pipeline.
-// It waits for the reset to occur. It simply unblocks the caller rather
-// than fully cancelling the reset request upon a context cancellation.
-func (s *Driver) ResetDerivationPipeline(ctx context.Context) error {
-	respCh := make(chan struct{}, 1)
-	select {
-	case <-ctx.Done():
-		return ctx.Err()
-	case s.forceReset <- respCh:
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		case <-respCh:
-			return nil
-		}
-	}
-}
-
-func (s *Driver) StartSequencer(ctx context.Context, blockHash common.Hash) error {
-	return s.sequencer.Start(ctx, blockHash)
-}
-
-func (s *Driver) StopSequencer(ctx context.Context) (common.Hash, error) {
-	return s.sequencer.Stop(ctx)
-}
-
-func (s *Driver) SequencerActive(ctx context.Context) (bool, error) {
-	return s.sequencer.Active(), nil
-}
-
-func (s *Driver) OverrideLeader(ctx context.Context) error {
-	return s.sequencer.OverrideLeader(ctx)
-}
-
-func (s *Driver) ConductorEnabled(ctx context.Context) (bool, error) {
-	return s.sequencer.ConductorEnabled(ctx), nil
-}
-
-func (s *Driver) SetRecoverMode(ctx context.Context, mode bool) error {
-	s.sequencer.SetRecoverMode(mode)
-	return nil
-}
-
-// SyncStatus blocks the driver event loop and captures the syncing status.
-func (s *Driver) SyncStatus(ctx context.Context) (*eth.SyncStatus, error) {
-	return s.StatusTracker.SyncStatus(), nil
-}
-
-// BlockRefWithStatus blocks the driver event loop and captures the syncing status,
-// along with an L2 block reference by number consistent with that same status.
-// If the event loop is too busy and the context expires, a context error is returned.
-func (s *Driver) BlockRefWithStatus(ctx context.Context, num uint64) (eth.L2BlockRef, *eth.SyncStatus, error) {
-	resp := s.StatusTracker.SyncStatus()
-	if resp.FinalizedL2.Number >= num { // If finalized, we are certain it does not reorg, and don't have to lock.
-		ref, err := s.L2.L2BlockRefByNumber(ctx, num)
-		return ref, resp, err
-	}
-	wait := make(chan struct{})
-	select {
-	case s.stateReq <- wait:
-		resp := s.StatusTracker.SyncStatus()
-		ref, err := s.L2.L2BlockRefByNumber(ctx, num)
-		<-wait
-		return ref, resp, err
-	case <-ctx.Done():
-		return eth.L2BlockRef{}, nil, ctx.Err()
-	}
-}
-
-// checkForGapInUnsafeQueue checks if there is a gap in the unsafe queue and attempts to retrieve the missing payloads from an alt-sync method.
-// WARNING: This is only an outgoing signal, the blocks are not guaranteed to be retrieved.
-// Results are received through OnUnsafeL2Payload.
-func (s *Driver) checkForGapInUnsafeQueue(ctx context.Context) error {
-	start := s.Engine.UnsafeL2Head()
-	end := s.CLSync.LowestQueuedUnsafeBlock()
-	// Check if we have missing blocks between the start and end. Request them if we do.
-	if end == (eth.L2BlockRef{}) {
-		s.log.Debug("requesting sync with open-end range", "start", start)
-		return s.altSync.RequestL2Range(ctx, start, eth.L2BlockRef{})
-	} else if end.Number > start.Number+1 {
-		s.log.Debug("requesting missing unsafe L2 block range", "start", start, "end", end, "size", end.Number-start.Number)
-		return s.altSync.RequestL2Range(ctx, start, end)
-	}
-	return nil
 }
